requirements.txt
flask
PyJWT
flask-cors
